Rename rate-limit error module to camelCase
Convex rejects module paths containing hyphens (must be alphanumeric,
underscores, or periods), which broke the deploy. Match the camelCase
convention used by the other files in the convex directory.
